+++ NEWS.md
@@ -2,11 +2,12 @@
 * `rSW2metrics` can now handle simulation output that is stored either in
   zipped archives (or in folders as before); metric and extraction functions
   gained a logical argument `zipped_runs`.
-* `metric_SW2toTable_daily` now correctly calculates
+* `metric_SW2toTable_daily()` now correctly calculates
   bulk volumetric water content (from soil moisture content) for the case when
   volumetric water content was not stored in output object.
-* `metric_SW2toTable_daily` now outputs daily potential evapotranspiration
-  as part of the variable group `"evapotranspiration"`.
+* `metric_SW2toTable_daily()` now provides additional outputs
+    * daily potential evapotranspiration (group `"evapotranspiration"`)
+    * daily total, wet, and dry degree days (group `"MDD"`)
 * `metric_RR2022predictors_annualClim()` now correctly calculates
   `"DeepDrainage_mean"` and `"CWD_mon_corr_temp_mean"`.
 * New `metric_RR2022predictors_annual()` provides annual time series of